Added [3Sum]

+++ Questions/LongestCommonPrefix/LongestCommonPrefix.cpp
@@ -7,17 +7,17 @@ public:
 		if(strs.size() == 0)
 			return sRet;
 
-		int nShortestLen = strs[0].size();
-		for(int i = 1; i < strs.size(); ++i)
+		size_t nShortestLen = strs[0].size();
+		for(size_t i = 1; i < strs.size(); ++i)
 		{
 			if(strs[i].size() < nShortestLen)
 				nShortestLen = strs[i].size();
 		}
 
-		for(int i = 0; i < nShortestLen; ++i)
+		for(size_t i = 0; i < nShortestLen; ++i)
 		{
 			char cCommon = strs[0].at(i);
-			for(int j = 1; j < strs.size(); ++j)
+			for(size_t j = 1; j < strs.size(); ++j)
 			{
 				if(strs[j].at(i) != cCommon)
 					return strs[0].substr(0, i);
